requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
pyfiglet==0.8.post1
python-magic==0.4.27
requests==2.29.0
soupsieve==2.4.1
urllib3==1.26.15
